pyproject.toml
[tool.poetry]
name = "disc-pal"
version = "0.1.0"
description = ""
authors = ["Ken Yoshizoe <kenyoshizoe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
